+++ keras_lmu/layers.py
@@ -101,11 +101,11 @@ class LMUCell(DropoutRNNCellMixin, tf.keras.layers.Layer):
         self.B = None
 
         if self.hidden_cell is None:
-            # if input_to_hidden=True then we can't determine the output size
-            # until build time
-            self.hidden_output_size = (
-                None if input_to_hidden else self.memory_d * self.order
-            )
+            for conn in ("hidden_to_memory", "input_to_hidden"):
+                if getattr(self, conn):
+                    raise ValueError(f"{conn} must be False if hidden_cell is None")
+
+            self.hidden_output_size = self.memory_d * self.order
             self.hidden_state_size = []
         elif hasattr(self.hidden_cell, "state_size"):
             self.hidden_output_size = self.hidden_cell.output_size
@@ -141,10 +141,6 @@ class LMUCell(DropoutRNNCellMixin, tf.keras.layers.Layer):
         """
 
         super().build(input_shape)
-
-        if self.input_to_hidden and self.hidden_cell is None:
-            self.hidden_output_size = self.memory_d * self.order + input_shape[-1]
-            self.output_size = self.hidden_output_size
 
         enc_d = input_shape[-1]
         if self.hidden_to_memory:
@@ -526,7 +522,12 @@ class LMUFFT(tf.keras.layers.Layer):
         if memory_d != 1:
             # TODO: we can support this by reusing the same impulse response
             #  for each dimension
-            raise NotImplementedError("Multi-dimensional memory not supported")
+            raise NotImplementedError(
+                "Multi-dimensional memory not supported in LMUFFT"
+            )
+
+        if input_to_hidden and hidden_cell is None:
+            raise ValueError("input_to_hidden must be False if hidden_cell is None")
 
         self.memory_d = memory_d
         self.order = order